basic: Have limit and counter to check the poll() behavior
It doesn't behave as I assumed, which is poll() will be called
multiple times until it receives futures::Ready(), as explained
in [getting asynchronous], but it's only called once and it
doesn't move forward.  I believe that's just because of
basic modules futures implementation but I couldn't figure that
out at this point.  It's TBD.

https://tokio.rs/docs/futures/getting_asynchronous/