echo: Dissy chain the Futures
I think there should be a better way to do this but for now, just
make it up and running.

But with this server, the clients are just crashing.  I've just
disable the client execution through main() for now.  I'll figure
it out once I write the client example based on the following
doc.

https://tokio.rs/docs/io/reading_writing_data/

Telnetting works, though.